backend/requirements.txt
blinker==1.9.0
requests>=2.31.0
click==8.1.8
colorama==0.4.6
Flask==3.1.3
flask-cors==6.0.2
bcrypt
google-genai
psycopg2-binary
PyJWT
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
Werkzeug==3.1.6
zipp==3.23.0
certifi>=2024.0.0
